【title】各类型订单页面以及功能完善
【description】各类型订单页面以及功能完善

【coder】coderXY

+++ swiftDeer/mainMoudle/orderModule/views/XYDeerOrderView.swift
@@ -9,7 +9,7 @@ import UIKit
 /** XYDeerOrderViewDelegate */
 @objc protocol XYDeerOrderViewDelegate {
     /** did selected index */
-    @objc func deerOrderHomeSelectedIdx(selectedIdx:NSInteger)
+    @objc func deerOrderHomeSelectedIdx(selectedIdx: Int)
 }
 
 class XYDeerOrderView: XYBaseView, UIScrollViewDelegate {
